Rename vars.py to consts.py
