+++ konsave/consts.py
